Disambiguates merged node names in collapsed tree
When two nodes are merged, the taxonomy term is used for naming it.
If there are more than one merged nodes with the same taxonomy term
in the collapsed tree, it is needed to rename them in order to
disambiguate. This commit adds an autoincremental integer in all
nodes where a taxonomy term is used in order to ensute that they
are always unique.